+++ src/components/Upload/Upload.scss
@@ -1,7 +1,6 @@
 .uploadpage-wrapper {
     width: 100vw;
     background-color: #f2f2f2;
-  }
 
   .uploadpage {
     margin: auto;
@@ -10,57 +9,4 @@
     margin-top: 46px;
     height: calc(100vh - 46px);
   }
-
-  .upload-box {
-    width: 800px;
-    height: 400px;
-    border: 1px solid #f2f2f2;
-    margin: auto;
-    display: flex;
-    text-align: center;
-    justify-content: center;
-    align-items: center;
-    margin-top: 150px;
-    box-shadow: rgba(0, 0, 0, 0.1) 0px 2px 12px -5px;
-
-    .upload-text {
-      font-size: 22px;
-      margin-bottom: 20px;
-    }
-
-    button {
-      width: 300px;
-      height: 40px;
-      background-color: #ff5500;
-      border: 1px solid #ff5500;
-      color: white;
-      font-size: 16px;
-      line-height: 40px;
-      margin-bottom: 10px;
-    }
-
-    .file-input {
-      display: none;
-    }
-
-    .upload-playlist {
-      display: flex;
-      text-align: center;
-      justify-content: center;
-    }
-
-    .upload-privacy {
-      display: flex;
-      text-align: center;
-      justify-content: center;
-      margin-top: 10px;
-
-      text{
-        margin-right: 10px;
-      }
-
-      .form-check {
-        margin-right: 10px;
-      }
-    }
-  }
+}
