+++ src/components/Upload/UploadModal/UploadModal.tsx
@@ -1,0 +1,16 @@
+import "./UploadModal.scss";
+
+function UploadModal() {
+  return (
+    <div className="upload-modal">
+      <div className="upload-modal-header">
+        <div className="upload-basic-info">Basic info</div>
+        <div>Metadata</div>
+        <div>Permissions</div>
+        <div>Advanced</div>
+      </div>
+    </div>
+  );
+}
+
+export default UploadModal;
